[chore] #141 Int64 변경

+++ Nearby/Nearby/Data/Network/Matching/DTO/MatchedCompanionListResponseDTO.swift
@@ -35,7 +35,7 @@ enum MatchStatus: String, Decodable {
 }
 
 struct MatchMyScheduleResponseDTO: Decodable {
-    let matchId: Int64
+    let matchId: Int
     let matchStatus: MatchStatus
     let schedule: Schedule?
     let openChatUrl: String?
